Started restructuring of docstring parser and generator.
