+++ pyment/__init__.py
@@ -1,5 +1,7 @@
-from .pyment import PyComment, __version__, __copyright__, __author__, __licence__
+"""Main pyment module."""
 
-name = "pyment"
+from .pyment import PyComment, __author__, __copyright__, __licence__, __version__
 
-__all__ = ['PyComment', '__version__', '__copyright__', '__author__', '__licence__']
+name = "pyment" # pylint: disable=invalid-name
+
+__all__ = ["PyComment", "__version__", "__copyright__", "__author__", "__licence__"]
