+++ pyment/docstringssssss/generators/__init__.py
@@ -1,0 +1,1 @@
+"""Module for generating docstrings."""
