+++ pyment/docstringssssss/parsers/__init__.py
@@ -1,0 +1,6 @@
+"""Module for parsing docstrings."""
+
+from .base_parser import DocToolsBase
+from .google_parser import GoogledocTools
+from .manager import Docs, DocsTools
+from .numpy_parser import NumpydocTools
