+++ pyproject.toml
@@ -85,7 +85,7 @@ dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 target-version = "py38"
 
 [tool.ruff.per-file-ignores]
-# "__init__.py" = ["E402", "F401"]
+"__init__.py" = ["E402", "F401"]
 "tests/test_*.py" = ["S101", "SLF001", "ANN101", "ANN102", "PLR2004"]
 
 [tool.ruff.pydocstyle]
@@ -208,7 +208,7 @@ evaluation = "max(0, 0 if fatal else 10.0 - ((float(5 * error + warning + refact
 
 [tool.pylint.similarities]
 # Minimum lines number of a similarity.
-min-similarity-lines = 4
+min-similarity-lines = 10
 
 [tool.pylint.spelling]
 # Limits count of emitted suggestions for spelling mistakes.
